requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.7.0
anyio==4.4.0
appdirs==1.4.4
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
attrs==23.2.0
Babel==2.15.0
beautifulsoup4==4.12.3
bitsandbytes==0.42.0
bleach==6.1.0
blinker==1.8.2
bokeh==3.4.2
bottle==0.12.25
bs4==0.0.2
cachetools==5.3.3
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorcet==3.1.0
comm==0.2.2
contourpy==1.2.1
cssselect==1.2.0
cycler==0.12.1
dataclasses-json==0.6.7
debugpy==1.8.2
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
executing==2.0.1
fake-useragent==1.5.1
fastjsonschema==2.20.0
feedparser==6.0.11
filelock==3.14.0
Flask==3.0.3
Flask-Cors==4.0.1
fonttools==4.53.0
fqdn==1.5.1
frozendict==2.4.4
frozenlist==1.4.1
fsspec==2024.6.0
gitdb==4.0.11
GitPython==3.1.43
google_search_results==2.4.2
h11==0.14.0
hf_transfer==0.1.6
holoviews==1.19.0
html5lib==1.1
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.23.3
hvplot==0.10.0
idna==3.7
importlib_metadata==8.0.0
ipykernel==6.29.4
ipympl==0.9.4
ipython==8.26.0
ipython-genutils==0.2.0
ipywidgets==8.1.3
ipywidgets_bokeh==1.6.0
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
json5==0.9.25
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
langchain==0.2.3
langchain-community==0.2.4
langchain-core==0.2.5
langchain-huggingface==0.0.3
langchain-openai==0.1.8
langchain-text-splitters==0.2.1
langchainhub==0.1.20
langgraph==0.0.66
langsmith==0.1.76
linkify-it-py==2.0.3
lxml==5.2.2
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
matplotlib==3.9.0
matplotlib-inline==0.1.7
mdit-py-plugins==0.4.1
mdurl==0.1.2
mistune==3.0.2
mlx==0.15.1
mlx-lm==0.14.3
mplchart==0.0.9
mplfinance==0.12.10b0
mpmath==1.3.0
multidict==6.0.5
multitasking==0.0.11
mypy-extensions==1.0.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
notebook==7.2.1
notebook_shim==0.2.4
numexpr==2.10.0
numpy==1.26.4
openai==1.33.0
orjson==3.10.4
overrides==7.7.0
packaging==23.2
pandas==2.2.2
pandocfilters==1.5.1
panel==1.4.4
param==2.1.1
parse==1.20.2
parso==0.8.4
peewee==3.17.5
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.2
prometheus_client==0.20.0
prompt_toolkit==3.0.47
protobuf==5.27.1
proxy_tools==0.1.0
psutil==6.0.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==16.1.0
pycparser==2.22
pydantic==2.7.3
pydantic_core==2.18.4
pydeck==0.9.1
pyee==11.1.0
Pygments==2.18.0
pyobjc-core==10.3.1
pyobjc-framework-Cocoa==10.3.1
pyobjc-framework-Security==10.3.1
pyobjc-framework-WebKit==10.3.1
pyparsing==3.1.2
pyppeteer==2.0.0
PyQt6==6.7.0
PyQt6-Qt6==6.7.2
PyQt6-sip==13.6.0
pyquery==2.0.0
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
pytz==2024.1
pyviz_comms==3.0.2
pywebview==5.1
PyYAML==6.0.1
pyzmq==26.0.3
qtconsole==5.5.2
QtPy==2.4.1
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
requests-html==0.10.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.18.1
safetensors==0.4.3
scikit-learn==1.5.0
scipy==1.13.1
Send2Trash==1.8.3
sentence-transformers==3.0.1
sentencepiece==0.2.0
setuptools==70.1.1
sgmllib3k==1.0.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.30
stack-data==0.6.3
streamlit==1.36.0
sympy==1.12.1
tavily-python==0.3.3
tenacity==8.3.0
terminado==0.18.1
text-generation==0.7.0
threadpoolctl==3.5.0
tiktoken==0.7.0
tinycss2==1.3.0
tokenizers==0.19.1
toml==0.10.2
toolz==0.12.1
torch==2.3.1
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
transformers==4.41.2
types-python-dateutil==2.9.0.20240316
types-requests==2.32.0.20240602
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
uc-micro-py==1.0.3
uri-template==1.3.0
urllib3==1.26.19
w3lib==2.2.1
watchfiles==0.22.0
wcwidth==0.2.13
webcolors==24.6.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==10.4
Werkzeug==3.0.3
widgetsnbextension==4.0.11
xyzservices==2024.6.0
yarl==1.9.4
yfinance==0.2.40
zipp==3.19.2
